finished hw 5

+++ hw5/gaussian_newton_ws/src/gaussian_newton_scanmatcher/src/map_cspace.cpp
@@ -118,8 +118,8 @@ void enqueue(map_t* map, unsigned int i, unsigned int j,
     return;
 
   //这里的距离是栅格的距离
-  unsigned int di = abs(i - src_i);
-  unsigned int dj = abs(j - src_j);
+  unsigned int di = abs((int)(i - src_i));
+  unsigned int dj = abs((int)(j - src_j));
   double distance = cdm->distances_[di][dj];
 
   if(distance > cdm->cell_radius_)
